Create my_printf.c
This is how I implement the printf() function in C.  The most valuable thing during this code was learning to struct the code, divide into smaller pieces, and deeper understanding of argv and argc arguments. Also, it gives me an understanding of how the command line interacts with the code. 
P.S.: The code could be optimized but in this manner, it is more clear for educational purposes.